Move FindUtils code in FindUtils.cc
Also remove static inline specifier. Most of them don't make sense since
the most functions are recursive.
